backend/configs/requirements/base.txt
python-slugify==8.0.4
Pillow==11.1.0
whitenoise==6.8.2
redis==5.2.1
hiredis==3.1.0
celery==5.4.0
django-celery-beat==2.7.0

django==5.0.10
django-compressor==4.5.1
django-redis==5.4.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt
requests
django-environ
pyotp
django-filter